requirements.txt
pandas
Pillow
torch
torchvision
transformers
numpy